confirm Delete modal ok

+++ members_handler_app/src/App.js
@@ -5,6 +5,7 @@ import Users from "./containers/Users";
 import Header from "./containers/Header";
 /*import SideBar from "./containers/SideBar";*/
 import Footer from "./containers/Footer";
+import Modal from "./components/Modal";
 
 export default class App extends Component {
   constructor(props) {
@@ -14,74 +15,27 @@ export default class App extends Component {
       inputValue: "",
       userId: "",
       uniqUser: { name: "", id: "" },
+      deleteModal: false,
     };
   }
 
-  handleInputValue(e) {
-    this.setState({ inputValue: e.target.value });
-  }
-
-  handleIdValue(e) {
-    this.setState({ userId: e.target.value });
-  }
-
-  handleSubmit(e) {
+  deleteContact = (e) => {
     e.preventDefault();
-    const data = { name: this.state.inputValue };
-    console.log("data", data);
+    let id = e.target.attributes.id.value;
+    this.setState({
+      deleteModal: !this.state.deleteModal,
+      uniqUser: { name: this.state.userList[id].name, id: id },
+    });
+  };
 
-    fetch("http://dry-forest-32366.herokuapp.com/api/users", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(data),
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        console.log("Success:", data);
-      })
-      .catch((error) => {
-        console.error("Error:", error);
-      });
-    this.setState({ inputValue: "" });
-  }
-
-  handleClick() {
-    axios
-      .get("http://dry-forest-32366.herokuapp.com/api/users")
-      .then((response) => this.setState({ userList: [...response.data] }));
-  }
-
-  handleModify(e) {
+  confirmDelete = (e) => {
     e.preventDefault();
-
-    const data = { name: this.state.inputValue };
-    fetch(
-      `http://dry-forest-32366.herokuapp.com/api/users/${this.state.userId}`,
-      {
-        method: "PUT",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(data),
-      }
-    )
-      .then((response) => response.json())
-      .then((data) => {
-        console.log("Success:", data);
-      })
-      .catch((error) => {
-        console.error("Error:", error);
-      });
-    this.setState({ inputValue: "", userId: "" });
-  }
-
-  handleDelete(e) {
-    e.preventDefault();
+    console.log(e);
+    console.log("deleted");
+    console.log(this.state.uniqUser.name, " ", this.state.uniqUser.id);
 
     fetch(
-      `http://dry-forest-32366.herokuapp.com/api/users/${this.state.userId}`,
+      `http://dry-forest-32366.herokuapp.com/api/users/${this.state.uniqUser.id}`,
       {
         method: "DELETE",
         headers: {
@@ -90,54 +44,151 @@ export default class App extends Component {
       }
     )
       .then((response) => JSON.stringify(response))
-      .then((data) => {
-        console.log("Success:", data);
+      .then(() => {
+        this.handleClick();
       })
       .catch((error) => {
         console.error("Error:", error);
       });
-    this.setState({ inputValue: "", userId: "" });
+    this.setState({
+      inputValue: "",
+      userId: "",
+      deleteModal: false,
+      uniqUser: { name: "", id: "" },
+    });
+  };
+
+  // handleInputValue(e) {
+  //   this.setState({ inputValue: e.target.value });
+  // }
+
+  // handleIdValue(e) {
+  //   console.log(e)
+  //   this.setState({ userId: e.target.value });
+  // }
+
+  // handleSubmit(e) {
+  //   e.preventDefault();
+  //   const data = { name: this.state.inputValue };
+  //   console.log("data", data);
+
+  //   fetch("http://dry-forest-32366.herokuapp.com/api/users", {
+  //     method: "POST",
+  //     headers: {
+  //       "Content-Type": "application/json",
+  //     },
+  //     body: JSON.stringify(data),
+  //   })
+  //     .then((response) => response.json())
+  //     .then((data) => {
+  //       console.log("Success:", data);
+  //     })
+  //     .catch((error) => {
+  //       console.error("Error:", error);
+  //     });
+  //   this.setState({ inputValue: "" });
+  // }
+
+  handleClick() {
+    axios
+      .get("http://dry-forest-32366.herokuapp.com/api/users")
+      .then((response) => this.setState({ userList: [...response.data] }));
   }
 
-  userClick(e) {
-    e.preventDefault();
-    this.setState({ userId: e.target.parentElement.id });
+  // handleModify(e) {
+  //   e.preventDefault();
 
-    fetch(
-      `http://dry-forest-32366.herokuapp.com/api/users/${e.target.parentElement.id}`,
-      {
-        method: "GET",
-        mode: "cors",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      }
-    )
-      .then((response) => response.json())
-      .then((data) => {
-        // console.log("Success:", data);
-        this.setState({
-          uniqUser: { name: data.name, id: data.id },
-        });
-      })
-      .catch((error) => {
-        console.error("Error:", error);
-      });
-    this.setState({ inputValue: "", userId: "" });
-  }
+  //   const data = { name: this.state.inputValue };
+  //   fetch(
+  //     `http://dry-forest-32366.herokuapp.com/api/users/${this.state.userId}`,
+  //     {
+  //       method: "PUT",
+  //       headers: {
+  //         "Content-Type": "application/json",
+  //       },
+  //       body: JSON.stringify(data),
+  //     }
+  //   )
+  //     .then((response) => response.json())
+  //     .then((data) => {
+  //       console.log("Success:", data);
+  //     })
+  //     .catch((error) => {
+  //       console.error("Error:", error);
+  //     });
+  //   this.setState({ inputValue: "", userId: "" });
+  // }
+
+  // handleDelete(e) {
+  //   e.preventDefault();
+
+  //   fetch(
+  //     `http://dry-forest-32366.herokuapp.com/api/users/${this.state.userId}`,
+  //     {
+  //       method: "DELETE",
+  //       headers: {
+  //         "Content-Type": "application/json",
+  //       },
+  //     }
+  //   )
+  //     .then((response) => JSON.stringify(response))
+  //     .then((data) => {
+  //       console.log("Success:", data);
+  //     })
+  //     .catch((error) => {
+  //       console.error("Error:", error);
+  //     });
+  //   this.setState({ inputValue: "", userId: "" });
+  // }
+
+  // userClick(e) {
+  //   e.preventDefault();
+  //   this.setState({ userId: e.target.parentElement.id });
+
+  //   fetch(
+  //     `http://dry-forest-32366.herokuapp.com/api/users/${e.target.parentElement.id}`,
+  //     {
+  //       method: "GET",
+  //       mode: "cors",
+  //       headers: {
+  //         "Content-Type": "application/json",
+  //       },
+  //     }
+  //   )
+  //     .then((response) => response.json())
+  //     .then((data) => {
+  //       // console.log("Success:", data);
+  //       this.setState({
+  //         uniqUser: { name: data.name, id: data.id },
+  //       });
+  //     })
+  //     .catch((error) => {
+  //       console.error("Error:", error);
+  //     });
+  //   this.setState({ inputValue: "", userId: "" });
+  // }
 
   componentWillMount() {
     this.handleClick();
   }
 
-  componentDidMount() {}
-
   render() {
     return (
       <div className="App row m-2">
         <Header />
-        <Users onLoad={this.handleClick} userList={this.state.userList} />
-       {/*} <SideBar
+        {!this.state.deleteModal ? null : (
+          <Modal
+            user={this.state.uniqUser}
+            confirmDelete={() => this.confirmDelete}
+          />
+        )}
+        <Users
+          onLoad={this.handleClick}
+          userList={this.state.userList}
+          deleteContact={(e) => this.deleteContact}
+          deleteModal={this.state.deleteModal}
+        />
+        {/*} <SideBar
           onAdd={(e) => this.handleSubmit(e)}
           handleInputValue={(e) => this.handleInputValue(e)}
           handleIdValue={(e) => this.handleIdValue(e)}
